backend/requirements.txt
flask-cors
psycopg2-binary
